Handle creating unique table names
Table names need to be unique across the entire workbook apparently.
Previously, I thought they only needed to be unique per each worksheet
but that appears to be incorrect.

Some additional logic was necessary to make that possible. Now, the
table name defaults to the current worksheet name if no table name is
provided.

Also, forgot to document the TableName parameter in Add-Table.