Added PivotFunction parameter
Added the ability to specify the summary function Excel uses to display
values in a PivotTable.

Previously, the summary function was not specified, and Excel
interpreted it as Sum by default. The Sum function is not terribly
useful when dealing with String data.

Note that this changes default behavior by specifying the default
function as Count.